pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["GodSkomis <60714418+GodSkomis@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
pytest = "^8.2.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
pytest-asyncio = "^0.23.7"
aiobotocore = "^2.13.0"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
